+++ R/estimate_peak.R
@@ -89,13 +89,13 @@
 estimate_peak <- function(x, n = 100L, alpha = 0.05, first_only = TRUE, progress = TRUE) {
 
     if (!inherits(x, "incidence2"))
-        .stopf_argument("`%s` is not an 'incidence2' object", deparse(substitute(x)))
+        .stopf("`%s` is not an 'incidence2' object.", deparse(substitute(x)))
 
-    .assert_scalar_whole(n)
-    .assert_scalar_numeric_not_na(alpha)
+    assert_scalar_whole(n)
+    assert_scalar_numeric_not_na(alpha)
 
-    .assert_bool(first_only)
-    .assert_bool(progress)
+    assert_bool(first_only)
+    assert_bool(progress)
 
     # Needed for CRAN/data.table
     ..date_var <- bootstrap_peaks <- ..observed_peak <- observed_count <- . <- i.bootstrap_peaks <- NULL
@@ -111,10 +111,10 @@ estimate_peak <- function(x, n = 100L, alpha = 0.05, first_only = TRUE, progress
     observed_peak <- keep_peaks(x, first_only = first_only)
     setDT(observed_peak)
     if (!first_only)
-        observed_peak <- observed_peak[,lapply(.SD, list), by = c(grouping_variables, count_val)]
+        observed_peak <- observed_peak[, lapply(.SD, list), by = c(grouping_variables, count_val)]
 
     # Calculate peaks from bootstrapped data samples with optional progress bar
-    if (progress) {
+    if (interactive() && progress) {
         out <- vector("list", n)
         message("Estimating peaks from bootstrap samples:")
         pb <- utils::txtProgressBar(min = 0, max = n, style = 3)
@@ -135,7 +135,7 @@ estimate_peak <- function(x, n = 100L, alpha = 0.05, first_only = TRUE, progress
 
     # TODO - check this with Thibaut
     # specify probabilities (lower_ci, median, upper_ci)
-    probs <- c(alpha/2, 0.5, 1 - alpha/2)
+    probs <- c(alpha / 2, 0.5, 1 - alpha / 2)
 
     # group peaks by group_vars and count_var
     peaks <- out[, .(bootstrap_peaks = list(.SD)), by = grouping_variables]