+++ R/keep.R
@@ -56,13 +56,13 @@ NULL
 keep_first <- function(x, n, complete_dates = TRUE, ...) {
 
     if (!inherits(x, "incidence2"))
-        .stop_argument("`x` must be an incidence2 object.")
+        stop("`x` must be an incidence2 object.")
 
-    .assert_scalar_whole(n)
+    assert_scalar_whole(n)
     if (n <= 0L)
-        .stop_argument("`n` must be non-negative")
+        stop("`n` must be non-negative")
 
-    .assert_bool(complete_dates)
+    assert_bool(complete_dates)
     if (complete_dates)
         x <- complete_dates(x, ...)
 
@@ -74,6 +74,13 @@ keep_first <- function(x, n, complete_dates = TRUE, ...) {
     tmp___index <- NULL
 
     # convert to data.table and order by date
+    if ("tmp___index" %in% names(x)) {
+        .stop(
+            "`keep_first` does not work if a column is already named 'tmp___index'. ",
+            "Please rename and try again. ",
+            "If this is problematic, please raise an issue."
+        )
+    }
     tmp <- as.data.table(x)
     tmp[, tmp___index := .I]
     setorderv(tmp, get_date_index_name.incidence2(x), order = 1L)
@@ -91,13 +98,13 @@ keep_first <- function(x, n, complete_dates = TRUE, ...) {
 keep_last <- function(x, n, complete_dates = TRUE, ...) {
 
     if (!inherits(x, "incidence2"))
-        .stop_argument("`x` must be an incidence2 object.")
+        stop("`x` must be an incidence2 object.")
 
-    .assert_scalar_whole(n)
+    assert_scalar_whole(n)
     if (n <= 0L)
-        .stop_argument("`n` must be non-negative")
+        stop("`n` must be non-negative")
 
-    .assert_bool(complete_dates)
+    assert_bool(complete_dates)
     if (complete_dates)
         x <- complete_dates(x, ...)
 
@@ -109,6 +116,13 @@ keep_last <- function(x, n, complete_dates = TRUE, ...) {
     tmp___index <- NULL
 
     # convert to data.table and order by date
+    if ("tmp___index" %in% names(x)) {
+        .stop(
+            "`keep_last` does not work if a column is already named 'tmp___index'. ",
+            "Please rename and try again. ",
+            "If this is problematic, please raise an issue."
+        )
+    }
     tmp <- as.data.table(x)
     tmp[, tmp___index := .I]
     setorderv(tmp, get_date_index_name.incidence2(x))
@@ -126,13 +140,13 @@ keep_last <- function(x, n, complete_dates = TRUE, ...) {
 keep_peaks <- function(x, complete_dates = TRUE, first_only = FALSE, ...) {
 
     if (!inherits(x, "incidence2"))
-        .stop_argument("`x` must be an incidence2 object.")
+        stop("`x` must be an incidence2 object.")
 
-    .assert_bool(complete_dates)
+    assert_bool(complete_dates)
     if (complete_dates)
         x <- complete_dates(x, ...)
 
-    .assert_bool(first_only)
+    assert_bool(first_only)
 
     # pull out grouping variables
     groups <- get_group_names.incidence2(x)
@@ -146,10 +160,18 @@ keep_peaks <- function(x, complete_dates = TRUE, first_only = FALSE, ...) {
     # convert to data.table
     tmp <- as.data.table(x)
 
+    # TODO - do this better
     # avoid check warnings
     tmp___index <- NULL
 
     # order by count
+    if ("tmp___index" %in% names(x)) {
+        .stop(
+            "`keep_peaks` does not work if a column is already named 'tmp___index'. ",
+            "Please rename and try again. ",
+            "If this is problematic, please raise an issue."
+        )
+    }
     tmp[, tmp___index := .I]
     tmp <- tmp[,
                list(tmp___index = tmp___index[.SD == max(.SD)]),