+++ R/mutate.R
@@ -48,8 +48,8 @@ mutate.incidence2 <- function(
     .before = NULL,
     .after = NULL
 ) {
-    if(!missing(.by)) {
-        .stop_argument("`.by` argument is not used in `mutate.incidence2()` as the groupings are implicit.")
+    if (!missing(.by)) {
+        stop("`.by` argument is not used in `mutate.incidence2()` as the groupings are implicit.")
     }
 
     groupings <- c(get_count_variable_name(.data), get_group_names(.data))