+++ R/plot.R
@@ -144,55 +144,51 @@ plot.incidence2 <- function(
 ) {
 
     if (!requireNamespace("ggplot2", quietly = TRUE))
-        .stop_suggested("'ggplot2' is required for the incidence plot method but is not present.")
+        stop("'ggplot2' is required for the incidence plot method but is not present.")
 
     # type checking
-    if (!is.numeric(width) || length(width) != 1L)
-        .stop_argument("`width` must be a numeric scalar.")
+    assert_scalar_numeric(width)
 
+    # NOTE - we keep this like this for backwards compatibility.
+    #        if we had used NA_character_ as the default we could just use
+    #        ympes::assert_scalar_character directly
     if (!is.na(border_colour)) {
-        if (!is.character(border_colour) || length(border_colour) != 1L) {
-            .stop_argument("`border_colour` must be a scalar character or, NA.")
-        }
+        assert_scalar_character(border_colour)
     }
 
-    if (!is.character(na_colour) || length(na_colour) != 1L)
-        .stop_argument("`na_colour` must be a scalar character.")
+    assert_scalar_character(na_colour)
 
-    if (!is.numeric(alpha) || length(alpha) != 1L)
-        .stop_argument("`alpha` must be a numeric scalar.")
+    assert_scalar_numeric(alpha)
 
     if (!is.null(fill)) {
+        assert_scalar_character(fill)
         if (!is.character(fill) || length(fill) != 1L) {
-            .stop_argument("`fill` must be a scalar character.")
+            stop("`fill` must be a scalar character or NULL.")
         }
     }
 
     if (!is.null(title)) {
         if (!is.character(title) || length(title) != 1L) {
-            .stop_argument("`title` must be a scalar character.")
+            stop("`title` must be a scalar character or NULL.")
         }
     }
 
-    if (!is.numeric(angle) || length(angle) != 1L)
-        .stop_argument("`angle` must be a numeric scalar.")
+    assert_scalar_numeric(angle)
 
     if (!is.null(size)) {
         if (!is.numeric(size) || length(size) != 1L) {
-            .stop_argument("`size` must be a numeric scalar or NULL.")
+            stop("`size` must be a numeric scalar or NULL.")
         }
     }
 
     if (!is.null(nrow)) {
         if (!is.numeric(nrow) || length(nrow) != 1L) {
-            .stop_argument("`nrow` must be a numeric scalar or NULL.")
+            stop("`nrow` must be a numeric scalar or NULL.")
         }
         nrow <- as.integer(nrow)
     }
 
-    if (!is.numeric(n_breaks) || length(n_breaks) != 1L) {
-        .stop_argument("`n_breaks` must be a numeric scalar.")
-    }
+    assert_scalar_numeric(n_breaks)
     n_breaks <- as.integer(n_breaks)
 
     legend <- match.arg(legend)
@@ -209,7 +205,7 @@ plot.incidence2 <- function(
     x_axis<- get_date_index_name.incidence2(x)
 
     # TODO - temporary measure until I can think about this more
-    supported = c(
+    supported <- c(
         "Date",
         "grates_period",
         "grates_isoweek",
@@ -221,10 +217,10 @@ plot.incidence2 <- function(
     )
     dates <- .subset2(x, x_axis)
     if (!inherits(dates, supported)) {
-        .stop_argument(paste0(
+        .stop(
             "The `date_index` of `x` must be a <Date> or <grates> object to use the provided `plot` method. ",
             "Please raise an issue at https://github.com/reconverse/incidence2/issues If you would like another object <type> supported"
-        ))
+        )
     }
 
     # create fill palette
@@ -235,7 +231,7 @@ plot.incidence2 <- function(
         fill_var <- count_var
     } else {
         if (!fill_var %in% names(x)) {
-            .stop_argument("`fill` must be the name of a column in `x`.")
+            stop("`fill` must be the name of a column in `x`.")
         }
     }
     fill <- .subset2(x, fill_var)