+++ R/rbind.R
@@ -1,10 +1,10 @@
 #' @export
-rbind.incidence2 <- function (..., deparse.level = 1) {
+rbind.incidence2 <- function(..., deparse.level = 1) {
     dat <- list(...)
     first <- dat[[1L]]
-    compatible <- sapply(dat, inherits, "data.frame")
+    compatible <- vapply(dat, inherits, TRUE, what = "data.frame")
     if (!all(compatible))
-        .stop_argument("Incompatible inputs")
+        stop("Incompatible inputs")
     dat <- lapply(dat, as.data.frame)
     out <- do.call(rbind, dat)
     .incidence_reconstruct(out, first)