+++ R/regroup.R
@@ -36,20 +36,19 @@
 #'
 # -------------------------------------------------------------------------
 #' @export
-regroup <- function(x, groups = NULL){
-
+regroup <- function(x, groups = NULL) {
     if (!inherits(x, "incidence2"))
-        .stop_argument("`x` must be an <incidence2> object.")
+        .stop("`x` must be an <incidence2> object.")
 
     # group checks
     group_variables <- attr(x, "groups")
     if (is.null(groups)) {
         if (!length(group_variables))
             return(x)
-    } else if(!(is.character(groups) && length(groups) >= 1L)) {
-        .stop_argument("`groups` must be NULL or a character vector.")
+    } else if (!(is.character(groups) && length(groups) >= 1L)) {
+        .stop("`groups` must be NULL or a character vector.")
     } else if (!all(groups %in% group_variables)) {
-        .stop_argument("Not all variables from `groups` are groupings of `x`.")
+        .stop("Not all variables from `groups` are groupings of `x`.")
     }
 
     # rebuild incidence
@@ -57,8 +56,8 @@ regroup <- function(x, groups = NULL){
     count_variable <- attr(x, "count_variable")
     count_value <- attr(x, "count_value")
 
-    # this is a little hacky
-    count_names_to = "temp_name"
+    # choose name not in use
+    count_names_to <- "temp_name"
     while (count_names_to %in% names(x))
         count_names_to <- basename(tempfile())
 
@@ -118,10 +117,9 @@ regroup <- function(x, groups = NULL){
 #'
 # -------------------------------------------------------------------------
 #' @export
-regroup_ <- function(x, groups = NULL){
-
+regroup_ <- function(x, groups = NULL) {
     if (!inherits(x, "incidence2"))
-        .stop_argument("`x` must be an <incidence2> object.")
+        stop("`x` must be an <incidence2> object.")
 
     groups_expr <- rlang::enquo(groups)
     groups_position <- tidyselect::eval_select(groups_expr, data = x)
@@ -132,15 +130,15 @@ regroup_ <- function(x, groups = NULL){
         return(x)
 
     if (!all(groups %in% group_variables))
-        .stop_argument("Not all variables from `groups` are groupings of `x`.")
+        stop("Not all variables from `groups` are groupings of `x`.")
 
     # rebuild incidence
     date_variable <- attr(x, "date_index")
     count_variable <- attr(x, "count_variable")
     count_value <- attr(x, "count_value")
 
-    # this is a little hacky
-    count_names_to = "temp_name"
+    # choose name not in use
+    count_names_to <- "temp_name"
     while (count_names_to %in% names(x))
         count_names_to <- basename(tempfile())