+++ R/summarise.R
@@ -45,11 +45,11 @@ summarise.incidence2 <- function(
     .by,
     .groups
 ) {
-    if(!missing(.by))
-        .stop_argument("`.by` argument cannot be used in `summary.incidence2()` as the groupings are implicit.")
+    if (!missing(.by))
+        stop("`.by` argument cannot be used in `summary.incidence2()` as the groupings are implicit.")
 
     if(!missing(.groups))
-        .stop_argument("`.groups` argument cannot be not used with `summary.incidence2()`.")
+        stop("`.groups` argument cannot be used with `summary.incidence2()`.")
 
     groupings <- c(get_count_variable_name(.data), get_group_names(.data))
     .data <- tibble::as_tibble(.data)