+++ NEWS.md
@@ -1,3 +1,17 @@
+# incidence2 2.4.0
+
+- `incidence()` gains a `fill` parameter. This is passed to `complete_dates()`
+  so is only valid when the argument `complete_dates = TRUE`.
+
+- For convenience we now reexport `select()` from
+  [dplyr](https://cran.r-project.org/package=dplyr) as well as 
+  `unnest()` and `unpack()` from [tidyr](https://cran.r-project.org/package=tidyr).
+  
+- Alt text has been added to all vignette images.
+
+- Improved input checking for `keep_first()` and `keep_last()`.
+
+
 # incidence2 2.3.1
 
 - Temporarily remove the [ympes](https://cran.r-project.org/package=ympes)